refactor: split linked page component
